requirements.txt
absl-py==2.1.0
agate==1.7.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.5.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
awswrangler==3.8.0
babel==2.16.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
blis==0.7.11
boto3==1.34.88
boto3-stubs==1.35.22
botocore==1.34.88
botocore-stubs==1.35.22
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==3.0.0
dbt-athena-community @ git+https://github.com/equiphealth/dbt-athena.git@5ad88e2e00309ab64486769cdff375130e4677f9
dbt-core==1.6.18
dbt-extractor==0.4.1
dbt-semantic-interfaces==0.2.3
debugpy==1.8.7
decorator==5.1.1
deepeval==1.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docx2txt==0.8
ed25519==1.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
googleapis-common-protos==1.65.0
grpcio==1.63.0
h11==0.14.0
hologram==0.0.16
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.20
langchain-text-splitters==0.2.4
langcodes==3.4.0
langsmith==0.1.125
language_data==1.2.0
leather==0.4.0
llvmlite==0.43.0
Logbook==1.5.3
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mashumaro==3.8.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mistune==3.0.2
more-itertools==8.14.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-boto3-athena==1.35.0
mypy-boto3-glue==1.35.18
mypy-boto3-lakeformation==1.35.0
mypy-boto3-sts==1.35.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.24.3
openai==1.35.7
openai-whisper==20240930
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.11.2
pexpect==4.9.0
phonenumbers==8.13.45
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAthena==3.9.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.18
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.16
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.1.20
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rouge-score==0.1.2
rpds-py==0.20.0
s3transfer==0.10.2
selenium==4.27.1
Send2Trash==1.8.3
sentry-sdk==2.14.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.4.2
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.5
tiktoken==0.7.0
tinycss2==1.4.0
tldextract==5.1.2
Token==1.8.6
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.29.2
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
types-awscrt==0.21.5
types-python-dateutil==2.9.0.20241003
types-s3transfer==0.10.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.32.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
whisper==1.1.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
